[OGUI-948] Add new grpc mock service
