+++ Control/test/test-config.js
@@ -25,14 +25,14 @@ module.exports = {
     hostname: 'localhost',
     port: 9090,
     timeout: 10000,
-    label: 'control',
+    label: 'Control',
     package: 'o2control'
   },
   apricot: {
     hostname: 'localhost',
     port: 9090,
     timeout: 10000,
-    label: 'apricot',
+    label: 'Apricot',
     package: 'apricot'
   },
   kafka: {
